refactor: App 컴포넌트에서 캘린더 뷰 및 이벤트 아이콘 분리
- CalendarView 컴포넌트 분리 (주간/월간 뷰 렌더링 로직)
- EventStatusIcons 컴포넌트 분리 (알림 및 반복 아이콘 표시)
- App.tsx 코드 간소화 및 가독성 향상
- 상수 파일 import 경로 수정
